dev: implement mobx store for workspace settings
