+++ web/components/workspace/settings/index.ts
@@ -1,0 +1,1 @@
+export * from "./workspace-details";
